Fix investments done over multiple periods
When using investments with a horizon of investments that were to be overwritten by the following solves, those investments in the horizon were included as invested capacity in the next solve. This is now removed. However, an issue remains when this is used in conjunction with dispatch runs, since it depends on realized_periods (https://github.com/irena-flextool/flextool/issues/58).