Lifetime_method introduced with options for reinvest_automatic and reinvest_choice.
Reinvest_choice means that at the end of the lifetime, the asset (unit, connection or storage in node) will cease to exist in the model unless the model decides to invest again. Automatic reinvestment means that whenever model invests in new capacity, it will be kept in the model forever through automatic reinvestments (the investment annuity is calculated until the end of model horizon for these investments).